manually merged assetto branch conflicts

+++ CrewChiefEx/sim_info.py
@@ -1,26 +1,17 @@
 """
 Assetto Corsa shared memory for Python applications
-
 _ctypes.pyd must be somewhere in sys.path, because AC doesn't include all Python binaries.
-
 Usage. Let's say you have following folder structure::
-
     some_app
         DLLs
             _ctypes.pyd
         some_app.py
-
 some_app.py::
-
     import os
     import sys
     sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'DLLs'))
-
     from sim_info import info
-
     print(info.graphics.tyreCompound, info.physics.rpms, info.static.playerNick)
-
-
 Do whatever you want with this code!
 WBR, Rombik :)
 """
@@ -221,11 +212,11 @@ class acsVehicleInfo(ctypes.Structure):
         ('speedMS', c_float),
         ('speedMPH', c_float),
         ('speedKMH', c_float),
-        ('bestLapMS', c_float),
+        ('bestLapMS', c_int32),
         ('lapCount', c_int32),
         ('currentLapInvalid', c_int32),
-        ('currentLapTimeMS', c_float),
-        ('lastLapTimeMS', c_float),
+        ('currentLapTimeMS', c_int32),
+        ('lastLapTimeMS', c_int32),
         ('localAngularVelocity', vec3),
         ('localVelocity', vec3),
         ('speedTotal', vec3),
